Added Waves and Drops
press C to go to the next wave after enemies have been killed. I added patches that spin and can be collected. The C key is temporary, next I we should instead make it so that the game scene changes when the wave is over and you need to click a button to go to the next scene